requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
beautifulsoup4==4.12.2
behave==1.2.6
behave-django==1.4.0
certifi==2023.7.22
charset-normalizer==3.3.2
decorator==5.1.1
Django==4.2.7
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-rest-framework==0.1.0
django-test-plus==2.2.3
djangorestframework==3.14.0
exceptiongroup==1.1.3
executing==2.0.1
geographiclib==2.0
geopy==2.4.0
h11==0.14.0
idna==3.4
ipdb==0.13.13
ipython==8.17.2
jedi==0.19.1
matplotlib-inline==0.1.6
outcome==1.3.0.post0
parse==1.19.1
parse-type==0.6.2
parso==0.8.3
pexpect==4.8.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PySocks==1.7.1
pytz==2023.3.post1
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.13.0
trio==0.23.0
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.9
wsproto==1.2.0
